requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cryptography==3.4.6
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.3
django-bootstrap4==2.3.1
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-stdimage==5.2.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
ldap3==2.9
model-mommy==2.0.0
NEMO==3.10.0
Pillow==8.1.2
progressbar2==3.53.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
pywin32==225
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.4
whitenoise==5.2.0
